consider script_args config var for script exec
First idea was:
must be a list of strings, that will build script argv[1:]

But then:
revert script_args to be a string

The argv format I better like for security and clarity, but this does
not go well with the CLI use case. I'm relunctant to using positional
arguments for that purpose, for a few reasons.

Let the customer escape and give a --script_args="-c -v /stuff"
argstring as is, as trying to split for him would be worse.

Better ideas welcome, here :/