swap_init.py: Use resolvconf on FreeBSD
resolvconf(8) help to manage multiple DNS configuration.
Using it allow the client to not break his DNS configuration
if he start dhclient by hand, or use a VPN that update DNS config.
It will also allow use to announce ipv6 DNS with the RA. Before that
it's not possible as rtsold(8) will call resolvconf which will erase
the existing configuration.