resolvconf: fix symlink in configuration
Due to an error in our build process for the
system image outside of this codebase, we used
to set resolv.conf.d/origin and create a symlink
'tail' to this file. The resolvconf documentation
state that the file should be name 'original' and
the symlink in the builded image was still
uncorrect. We  recreate the resolv.conf.d/tail
symlink.

Signed-off-by: aegiap <aegiap@gandi.net>